Refactor pages
