+++ packages/catch-a-nest-frontend-ssr/pages/_app.tsx
@@ -1,71 +1,18 @@
 import React, { useState } from 'react';
 import { QueryClient, QueryClientProvider } from 'react-query';
 import { Hydrate } from 'react-query/hydration';
+import Head from 'next/head';
 import type { AppProps } from 'next/app';
 import { ThemeProvider } from 'next-themes';
-import Head from 'next/head';
 import { RecoilRoot } from 'recoil';
+import { darkThemeClassName, lightThemeClassName } from '@stitches.js';
 import RecoilInitializer from '@/components/RecoilInitializer';
 import FullscreenLoader from '@/components/FullscreenLoader';
-import { DefaultSEO } from '@/components/AppSEO/DefaultSEO';
-import AppToastProvider from '@/components/AppToastProvider';
-import FirebaseAnalytics from '@/components/FirebaseAnalytics';
-import DebugComponents from '@/components/DebugComponents';
-import {
-  darkThemeClassName,
-  globalCss,
-  lightThemeClassName,
-} from '@stitches.js';
-
-const globalStyle = globalCss({
-  '*, *::before, *::after': {
-    boxSizing: 'border-box',
-  },
-
-  'html, body, #__next': {
-    height: '100%',
-  },
-
-  body: {
-    m: 0,
-    fontFamily: '$base',
-    WebkitFontSmoothing: 'antialiased',
-    MozOsxFontSmoothing: 'grayscale',
-    WebkitTextSizeAdjust: '100%',
-
-    bc: '$loContrast',
-  },
-
-  button: {
-    border: 'none',
-    m: 0,
-    p: 0,
-    color: 'inherit',
-    font: 'inherit',
-    background: 'transparent',
-    WebkitFontSmoothing: 'inherit',
-    MozOsxFontSmoothing: 'inherit',
-    appearance: 'none',
-
-    cursor: 'pointer',
-  },
-
-  a: {
-    outline: 'none',
-    m: 0,
-    p: 0,
-    color: 'inherit',
-    font: 'inherit',
-    textDecoration: 'none',
-  },
-
-  'h1, h2, h3, h4, h5, h6': {
-    wordBreak: 'keep-all',
-  },
-  'p, span': {
-    wordBreak: 'keep-all',
-  },
-});
+import { DefaultSEO } from '@/components/SEO/DefaultSEO';
+import ToastProvider from '@/components/ToastProvider';
+import globalStyle from '@/lib/styles/globalStyle';
+import DebugComponents from '@/components/debug/DebugComponents';
+import FirebaseAnalytics from '@/components/analytics/FirebaseAnalytics';
 
 function MyApp({ Component, pageProps }: AppProps) {
   globalStyle();
@@ -96,7 +43,7 @@ function MyApp({ Component, pageProps }: AppProps) {
         <DefaultSEO />
         <RecoilRoot>
           <RecoilInitializer />
-          <AppToastProvider />
+          <ToastProvider />
           <QueryClientProvider client={queryClient}>
             <Hydrate state={pageProps.dehydratedState}>
               <Component {...pageProps} />
