+++ packages/catch-a-nest-frontend-ssr/pages/_document.tsx
@@ -32,6 +32,10 @@ class MyDocument extends Document {
             rel="stylesheet"
           />
           <link
+            href="https://fonts.googleapis.com/css2?family=Fira+Code:wght@500;700&display=swap"
+            rel="stylesheet"
+          />
+          <link
             rel="apple-touch-icon"
             sizes="180x180"
             href="/apple-touch-icon.png"
